Check file existence before unlinking it
Check whether a file exists before unlinking it. Unlinking a non-existing file will throw an error message. See ClarkSource/ember-makeup#242

+++ index.js
@@ -309,7 +309,7 @@ module.exports = class Filter extends Plugin {
         result = await this.processAndCacheFile(srcDir, destDir, entry, forceInvalidation, isChange, stats);
       } else {
         stats.linked++;
-        if (isChange) {
+        if (isChange && fs.existsSync(outputPath)) {
           fs.unlinkSync(outputPath);
         }
         result = symlinkOrCopySync(srcPath, outputPath);
